Redid the request for location services, added onStart
I had to redo the location service request because it works nothing like the bluetooth request at all and basically did nothing as it was. It now drops the user into the settings application in the right activity and gives them a toast what to do (turn on location settings), anything else is really complicated and involves using the Google API on which I'm not too keen. I also had to include a small onStart function because once the user enters the settings that means our Activity is no longer visible and therefore enters the stopped state and once the user goes back there simply was no control flow in our hands anymore.